requirements.txt
mss>=9.0.0
pynput>=1.7.6
Pillow>=10.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-docx>=1.0.0
